+++ src/richie/apps/search/indexers/organizations.py
@@ -34,7 +34,10 @@ class OrganizationsIndexer:
             # Searchable
             # description & title are handled by `MULTILINGUAL_TEXT`
             **{
-                "complete.{:s}".format(lang): {"type": "completion"}
+                "complete.{:s}".format(lang): {
+                    "type": "completion",
+                    "analyzer": "simple_diacritics_insensitive",
+                }
                 for lang, _ in settings.LANGUAGES
             },
             # Not searchable